frontend: move Configure GitHub button to user pill in TopBar
The Configure GitHub button was in the sidebar footer, which cluttered a task-centric panel. It now lives on the user avatar/login pill in the top bar — clicking your GitHub avatar opens the GitHub config modal, which is the standard UX pattern for account-level settings.
